Clarify licensing
Move each license to a better named file. Create LICENSE file describing
osquery is dual licensed and clarify the user can pick one of the two
licenses when using osquery. Add SPDX.

This will allow a user that opens the LICENSE file to immediately
understand the licensing of osquery which wasn't the case before.