Helper functions (tryTake, tryTakeCopy, getOr) to lookup in key-value tables
There are a lot of looups in the maps the osquery code. Most of them are verbose and not-optimal with check if such key exists in the table before get access. Some of them consists error e.g.:
```c++
r["uid"] = row.count("uuid") > 0 ? row.at("uid") : "";
```

Introduced code will help to avoid the most of such problems.