Create an abstraction for accessing columns of a table row
This is the core of the change -- making it possible to have different
ways of accessing the columns. The "legacy" way becomes `DynamicTableRow`,
backed by a `Row`. Later on we'll generate `TableRow` subclasses for each
table.

As of this commit, `tables.cpp` and `virtual_table.cpp` compile, but lots
of other stuff doesn't. The following few commits will fix up the rest of
the code.