events/windows: Remove empty old path check
Experimentally, lots of events are missing their old path,
and so we're dropping a lot of events that we should probably be
emitting.

We also have a problem in the form of precomputing the globs that
we match the path and old path against, which will fail to catch
cases like e.g. renaming a file to a new name in the same directory.