Change semantics of 'start_time' colume in processes table
Summary: Now for different platforms this columns used to mean different things. On posix it was seconds since system boot. On linux implementation was platoform specific and on some configurations it was seconds since boot, on others it was scaled seconds (x10, x100, x0.1 etc). On windows it was abs unix time in seconds since Epoch. In order to make it consistent for all OS we changed to more convinient format - absolute time since Epoch. This commit is about linux. Next one going to be about Darwin.

Differential Revision: D13918626

fbshipit-source-id: 1b17f4929af71acc4539806a6f8f2cb9f6c3ef41