Fix up start_time column for macos (#5412)
Summary:
Pull Request resolved: https://github.com/facebook/osquery/pull/5412

Now on different platforms column `start_time` in `processes` table means different things. On Linux it is seconds since system boot, but it works correct only for some platforms, because the number of clock ticks per second was hardcoded. On windows it was abs unix time in seconds since Epoch. On macos it is a time in milliseconds (may be?) since system boot. On freeBSD as far as I can see it an abs time since boot, but also I'm not sure.

In order to make it consistent for all OS we changed to more convenient format - absolute time since Epoch. This commit is about macos.

Reviewed By: marekcirkos

Differential Revision: D13918625

fbshipit-source-id: eacb297358b36ce72cb0d5a7d9171553688ab2a3