prototype of osquery/plugin_sdk
Summary:
Build osquery extensions/plugins against all osquery is an apparently bad idea.
Because plugins usually are not require full set of thirdparty libs of osquery. Most of osquery code is also not necesarry for plugins. And the main thing, clear definition of public interface of osquery plugins will force us to keep it clear and short, change it carefully and consciously, crate a tests for it.

Reviewed By: fmanco

Differential Revision: D13990668

fbshipit-source-id: ed6ed3f6f75178d829fc6bcbd1af20ef2e268fa8