Relicense yara_utils to Facebook
Summary: This file was originally written by wxsBSD in 2015. He has since joined Facebook and has graciously agreed to re-license this file to Facebook. This diff formalizes the relicensing by changing the copyright notice on the file. Note that wxsBSD still retains a copyright to all previous versions of the file.

Reviewed By: wxsBSD

Differential Revision: D14131447

fbshipit-source-id: 2fd5c30e9131f8c16bae9ee5ea280ba081806cdc