Relicense yara_utils to Facebook (#5450)
Summary:
Pull Request resolved: https://github.com/facebook/osquery/pull/5450

This file was originally written by wxsBSD in 2015. He has since joined Facebook and has graciously agreed to re-license this file to Facebook. This diff formalizes the relicensing by changing the copyright notice on the file. Note that wxsBSD still retains a copyright to all previous versions of the file.

Reviewed By: wxsBSD

Differential Revision: D14131447

fbshipit-source-id: 3148eafc0162a23b86e064a9784ea01b685164ef