`impl_fbthrift` injection for fbcode build of osquery PROD extensions
Summary:
Let's create a fbcode wrapper for osuery plugin SDK to wrap it up with fbthrift implementation from fbcode. Also we should trick the build system and add empty function `registerForeignTables()` to avoid including all osquery builtin tables implementation.

Since that, SDK suppose to be full functional and ready to be used to build facebook PROD extensions in fbcode with dependency on it (see next diff).

Reviewed By: fmanco

Differential Revision: D14224205

fbshipit-source-id: c08cbb374f2d68f6b93637809b97dcf0159c9e08