plugin SDK tests: registry should contain only optional plugins and some mandatory registries
Summary:
osquery registries is big part of SDK interface and it is important to make
sure SDK include them all.

Need of tests for non optional plugins can be not so obvious. All non-optional
plugins should not be included to SDK library. Because it causes to double
plugin registration when extensions process connects to `osqueryd`. That leads
to the crash of extensions process.

Unfortunatelly there is no explisit list of plugins included to SDK and there is
no way to check them at compile time. Thankfully it can be done at
runtime and we can make a test for it.

Differential Revision: D14261046

fbshipit-source-id: 1d02118ffd9994fbcd48464f6124fdfc7d32c5ba