Refactor to use generic lambda deleter
Rather than definitely multiple copies of a deleter function, that are
identical except for the pointer type, we can use a generic lambda that
is reusable for all of the smart pointer.

There is one wart: to initialize an empty (nullptr) `std::shared_ptr`
with a generic lambda deleter, one must cast the nullptr argument to
avoid a compiler error.