Introduce events stream registry
Summary: It is experimental registry to stream events from osquery. It is separated from logger, which is semantically very similar, on purpose. Firstly becaus we are still not sure about exact way to process/filter/deliver events, so I thing the best way to test things is commit is separatelly for now and decide later. It's easier to merge things than split them. Sec

Differential Revision: D14404665

fbshipit-source-id: 44ea74cdf6d479f4413deb900b94720868e60a95