Add more test cases to query platform unit test (#5515)
Summary:
Pull Request resolved: https://github.com/facebook/osquery/pull/5515

This diff augments the unit test which determines if a query should execute on this client. We add more test cases for different platforms, including the ones which are no longer officially supported, `centos` and `ubuntu`. These values remain equivalent to `linux`. That is, a query with either of those values specified for the platform will be executed on any Linux system.

Reviewed By: guliashvili

Differential Revision: D14430282

fbshipit-source-id: 64a47ded0ef769d927753b9d75a378b25a2391bd