Total(user+kernel) time ODS timeseries
Summary: Before, we sent the user and kernel time in a separate time series. They are usefull separately, in order to measure which queries are syscall extensive. However, for measuring the total time we need them together. We have wall time, but user + kernel time is more precise as they measure resource usage of the thread.

Reviewed By: jessek

Differential Revision: D14502715

fbshipit-source-id: 7a03ec5b1d7b328e9657d94f752d33e7aa771e97