schemer formatters: json serializer (#5545)
Summary:
Pull Request resolved: https://github.com/facebook/osquery/pull/5545

This is a JSON serializing formatter for schemer. It represents C++ object as JSON object according to defined in C++ class schema. The implementation based on rapidjson library. It is very simple - just print all key:value pairs directly to rapidjson stream (it could be files stream or string stream). The second method of formatter converts C++ objects directly to JSON in string.

Two methods with the same name: `osquery::schemer::toJson`

Reviewed By: SAlexandru

Differential Revision: D14663996

fbshipit-source-id: 7a902b0777c9094faf39ee3111bbc09d8c3c7a55