Handle bad fds in md_tables
Summary:
Infer flagged this use of `unique_ptr` as potentially leaking the FD. I do not think this is the case but there is a missing check for a failed `open`. I am unsure of what the `ioctl` would do in this case.

I removed the custom decl to make the logic in this table easier (opinion).

Reviewed By: guliashvili

Differential Revision: D14700412

fbshipit-source-id: 5747494356e4dac88df03bb7245e77e90c3ca43c