Handle bad fds in md_tables (#5553)
Summary:
Pull Request resolved: https://github.com/facebook/osquery/pull/5553

Infer flagged this use of `unique_ptr` as potentially leaking the FD. I do not think this is the case but there is a missing check for a failed `open`. I am unsure of what the `ioctl` would do in this case.

I removed the custom decl to make the logic in this table easier (opinion).

Reviewed By: guliashvili

Differential Revision: D14700412

fbshipit-source-id: cfcfe9d2629aa27f18d39058830cb5c63881f163