Struct [HostIdentity] to represent minimal info to identify a certain host in osquery
Summary:
I'm going to it as one of the decorators for streaming events. To be able to associate event with a certain machine on the backend side.

Why a new directory in osquery/? There are some libraries that depend on
core/database and any monster modules. To be able to keep utils simple,
lightweight and independent we should not put stuff like filesystem or network
under utils/ directory. But we need a single place to put all system assosiated
libs. So, let's make a `system` in the root of osquery.

Reviewed By: guliashvili

Differential Revision: D14706186

fbshipit-source-id: d40fde3872ca6b6677a1d8f89cfd8eda63c6b83d