deployment: updating osquery installation path to reside in Program Files
osquery as it stands is vulnerable to a privilege escalation attack, if a
malicious attacker is able to inject a new executable path into the extensions.load
file, and hard link a parent folder of a malicious binary to a folder with known
'safe' permissions, osquery will load said malicious executable with SYSTEM
permissions. This can be mitigated by migrating our installation to the
'Program Files' directory on windows which restricts unprivileged write access.