Fix prepare_for_ide target on macOS and Windows
This is a hotfix to let the prepare_for_ide target work on Windows and
macOS, avoiding to download libudev that it's not needed and
doesn't exist for those platforms.

In theory for all the third-party libraries that do not exist for a
platform we should have an INTERFACE library as a substitute,
instead of avoiding to define any target.
We wait to do this change, because we are going to refactor how
dependencies are downloaded and imported.