rpm: Check for glibc version and librpm secure_getenv
The librpm dependency may be configured with a glibc version < 2.17. If
so then it will expect __secure_getenv to be available.

cmake/prebuilt: Add definePrebuiltProject and update importThirdPartyLibrary

- New function definePrebuiltProject as a ExternalProject wrapper
- Update importThirdPartyLibrary (Binary and HeaderOnly) functions
- THIRD_PARTY_PREBUILT_PATH points to the submodule root

This change also reorders the third-party project list in
third-party/CMakeLists.txt so that recursive dependency projects are
defined.

cmake/prebuilt: Update third-party projects

This is a large refactor that updates the API usage for importing
third-party libraries, now they are discovered within the submodule vs.
downloaded and extracted.

The logic is ported from the old provision code for Homebrew and Choco.
This includes patch files (hence the update to .gitignore) and a
inline-replace python script to handle CMake/sed/perl oddities.

The ncurses library is introduced as a recursive dependency but it is
not used in the final build.

cmake/prebuilt: Add helper script to build on macOS and Linux

The helper ./tools/build_prebuilt.sh is helpful on Linux (and MacOS) to
assure a reproducible build.

cmake/prebuilt: Remove cloud tables from windows

The current prebuilt-build produces a large AWS-SDK-EC2 library that we
cannot host in the submodule. It does not look like the EC2 tables are
available on Windows (due to spec-file configuration).

Please see https://osquery.io/schema/3.3.2#ec2_instance_tags

This removes the EC2 library linking requirement until we revisit this
feature.

cmake/prebuilt: Add submodule and update documentation

- Add the third-party-prebuilt submodule
- Update azure pipelines to checkout the submodule
- Refactor the build documentation and include prebuilt info

cmake/prebuilt: Fix make macros and ninja compatibility

This commit applies some feedback and most importantly does not create
the build-from-source targets if the generator is Ninja. This is
because some ExternalProject commands use $(MAKE) and this breaks the
Ninja.build.

Azure: Enable CI on the thirdparty-submodules branch