cmake/prebuilt: Use a submodule to host prebuilt dependencies
There are several changes in this commit:

1. Add definePrebuiltProject and update importThirdPartyLibrary CMake functions.
   The definePrebuiltProject function is a wrapper for CMake's ExternalProject.
2. Define a CMake option THIRD_PARTY_PREBUILT_PATH, which points to the
   submodule root. This contains prefixes for each supported OS.
3. Add the third-party-prebuilt submodule to the root. Update Azure pipelines
   to checkout the submodule.
4. Refactor the build documentation and include prebuilt info.
5. Include Augeas lenses within the prebuilt prefixes.

This change also reorders the third-party project list in
third-party/CMakeLists.txt so that recursive dependency projects are
defined.

The build logic for each prebuilt dependency is ported from the old
provisioning code for Homebrew and Choco. This includes patch files
(hence the update to .gitignore) and a inline-replace python script to handle
CMake/sed/perl oddities. The ncurses library is introduced as a recursive
dependency but it is not used in the final build.

The helper ./tools/build_prebuilt.sh is helpful on Linux (and MacOS) to
assure a reproducible build.