cmake/prebuilt: Remove cloud tables from windows
The current prebuilt-build produces a large AWS-SDK-EC2 library that we
cannot host in the submodule. It does not look like the EC2 tables are
available on Windows (due to spec-file configuration).

Please see https://osquery.io/schema/3.3.2#ec2_instance_tags

This removes the EC2 library linking requirement until we revisit this
feature.