Make Personal certs visible for other users
They aren't returned by the cert enumeration APIs, but if we
directly look into their %APPDATA%, we can find them. The files
there are a custom file format, which we can parse to extract the
actual encoded certificate. We can then forward that back to the
winapi to extract info out of it.

We can also proactively look for them on disk, regardless of the
enumeration, which will guarantee that we will always be able to show
Personal certs regardless of things like, e.g. login history. For
example, if a user has never logged in on a system since boot, they may
not appear in the registry, and thus the enumeration, so their Personal
certs will not appear, even though they are available.